backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Caching==2.1.0
nba_api==1.2.1
pydantic==2.4.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
pandas==2.1.1
numpy==1.24.3